Development

+++ FrontEnd/src/Pages/Customers/index.js
@@ -5,7 +5,7 @@ import axios from 'axios';
 import SideMenu from '../../Components/SideMenu';
 import PageContent from '../../Components/PageContent';
 
-function Cutomer() {
+function Customer() {
 
     const [users,setUsers]=useState([]);
 
@@ -20,10 +20,23 @@ function Cutomer() {
         setUsers(result.data);
     };
 
-    const deleteUser = async (customer_id) => {
-        const result = await axios.delete(`http://localhost:5000/user/${user_id}`)
+    // const deleteUser = async (user_id) => {
+    //     const result = await axios.delete(`http://localhost:5000/user/${user_id}`)
+    //     loadUsers();
+    // }
+
+    const deleteUser = async (user_id) => {
+        // Show a confirmation popup
+        const confirmDelete = window.confirm("Are you sure you want to delete this customer?");
+        if (confirmDelete) {
+        // Delete the category
+        await axios.delete(`http://localhost:5000/user/${user_id}`);
         loadUsers();
-    }
+        } else {
+        // Do not delete the category
+        console.log("Customer was not deleted.");
+        }
+    };
 
     return (
         <div className="SideMenuAndPageContent">
@@ -31,18 +44,15 @@ function Cutomer() {
         <PageContent></PageContent>
             <div>
                 <Typography.Title level={4}>Our Users</Typography.Title>
-                {/* <Link className='btn btn-primary' to="/addnewcategory">Add New Customer</Link> */}
+                {/* <Link className='btn btn-primary' to="/addnewcustomer">Add New Customer</Link> */}
                 <div className="container">
                     <div className="py-4">
                         <table className="table border shadow-inner, table-primary">
                         <thead className="thead-dark">
                             <tr>
                                 <th scope="col">Customer ID</th>
-                                <th scope="col">Customer Telephone</th>
                                 <th scope="col">Customer First Name</th>
-                                <th scope="col">Customer Last Name</th>
-                                <th scope="col">Customer Address</th>
-                                <th scope="col">Customer City</th>
+                                <th scope="col">Customer Telephone</th>
                                 <th scope="col">Customer Email</th>
                                 <th scope="col">Action</th>
                             </tr>
@@ -52,21 +62,18 @@ function Cutomer() {
                             users.map((user,index)=>(
                                 <tr>
                                     <th scope="row" key={index}>{index+1}</th>
-                                    <td>{user.phoneno}</td>
                                     <td>{user.fname}</td>
-                                    <td>{user.lname}</td>
-                                    <td>{user.address}</td>
-                                    <td>{user.city}</td>
+                                    <td>{user.phoneno}</td>
                                     <td>{user.emailaddress}</td>
                                     <td>
                                         <Link className='btn btn-primary mx-2'
-                                            to={`/viewcategory/${user.userid}`}
+                                            to={`/viewcustomer/${user.userid}`}
                                         >
                                             View
                                         </Link>
 
                                         <Link className='btn btn-outline-primary mx-2'
-                                        to={`/editcategory/${user.userid}`}
+                                        to={`/editcustomer/${user.userid}`}
                                         >
                                             Edit
                                         </Link>
@@ -88,4 +95,4 @@ function Cutomer() {
         </div>
     ); 
 }
-export default Cutomer
+export default Customer
